got the add order working

+++ src/main/java/com/ships/repositories/ShippingCompanyInterface.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.ships.model.ShippingCompany;
 
 @Repository
-public interface ShippingCompanyInterface extends CrudRepository<ShippingCompany, Long>{
+public interface ShippingCompanyInterface extends CrudRepository<ShippingCompany, Integer>{
 
 }
